Added `readonly` property to `Post` component
`Post` component uses `readonly` property to determine whether or not it should allow user input on buttons like the `FlagButton` or the `VoteButton`.

UI design/implementation team (or frontend team I guess?) needs to add UI feedback to reflect this new `disabled` state for the `Post` buttons.

‼️ They also need to move `app/page.tsx` `Post` code to the actual `Post.tsx` component (as these two are currently out of sync and not used consistently). ‼️

P.s., on line 184 in `app/page.tsx`, should it be `/event/` or `/events/`?